09.14 : priority queue(가운데를 말해요)

+++ baekjoon/1655.js
@@ -36,7 +36,7 @@ class MaxHeap {
   };
 
   delete = () => {
-    if (!this.size) return;
+    if (this.size === 0) return;
     if (this.size === 1) this.#heap.pop();
     else this.#heap[ROOT] = this.#heap.pop();
     this.size--;
@@ -47,25 +47,18 @@ class MaxHeap {
     let isLeftBig = true;
     // 자식 노드가 하나도 없을 때까지
     while (this.#heap[left] !== undefined) {
-      // 오른쪽 자식 노드가 없는 경우
-      if (this.#heap[right] === undefined) {
-        if (this.#heap[curr] > this.#heap[left]) return;
-      }
       // 오른쪽 자식 노드가 있는 경우
-      else {
-        if (
-          this.#heap[curr] > this.#heap[left] &&
-          this.#heap[curr] > this.#heap[right]
-        )
-          return;
-        if (this.#heap[left] < this.#heap[right]) {
-          isLeftBig = false;
-        }
+      if (this.#heap[right] !== undefined) {
+        // 여기서 else 문을 주지 않아서 모든 경우의 수를 처리할 수 없었음
+        if (this.#heap[left] > this.#heap[right]) isLeftBig = true;
+        else isLeftBig = false;
       }
-      // swap
-      tmp = this.#heap[isLeftBig ? left : right];
-      this.#heap[isLeftBig ? left : right] = this.#heap[curr];
-      this.#heap[curr] = tmp;
+      if (this.#heap[curr] < this.#heap[isLeftBig ? left : right]) {
+        // swap
+        tmp = this.#heap[isLeftBig ? left : right];
+        this.#heap[isLeftBig ? left : right] = this.#heap[curr];
+        this.#heap[curr] = tmp;
+      }
       curr = isLeftBig ? left : right;
       left = curr * 2;
       right = left + 1;
@@ -106,7 +99,7 @@ class MinHeap {
   };
 
   delete = () => {
-    if (!this.size) return;
+    if (this.size === 0) return;
     if (this.size === 1) this.#heap.pop();
     else this.#heap[ROOT] = this.#heap.pop();
     this.size--;
@@ -117,25 +110,18 @@ class MinHeap {
     let isLeftSmall = true;
     // 자식 노드가 하나도 없을 때까지
     while (this.#heap[left] !== undefined) {
-      // 오른쪽 자식 노드가 없는 경우
-      if (this.#heap[right] === undefined) {
-        if (this.#heap[curr] < this.#heap[left]) return;
-      }
       // 오른쪽 자식 노드가 있는 경우
-      else {
-        if (
-          this.#heap[curr] < this.#heap[left] &&
-          this.#heap[curr] < this.#heap[right]
-        )
-          return;
-        if (this.#heap[left] > this.#heap[right]) {
-          isLeftSmall = false;
-        }
+      if (this.#heap[right] !== undefined) {
+        // 여기서 else 문을 주지 않아서 모든 경우의 수를 처리할 수 없었음
+        if (this.#heap[left] < this.#heap[right]) isLeftSmall = true;
+        else isLeftSmall = false;
       }
-      // swap
-      tmp = this.#heap[isLeftSmall ? left : right];
-      this.#heap[isLeftSmall ? left : right] = this.#heap[curr];
-      this.#heap[curr] = tmp;
+      if (this.#heap[curr] > this.#heap[isLeftSmall ? left : right]) {
+        // swap
+        tmp = this.#heap[isLeftSmall ? left : right];
+        this.#heap[isLeftSmall ? left : right] = this.#heap[curr];
+        this.#heap[curr] = tmp;
+      }
       curr = isLeftSmall ? left : right;
       left = curr * 2;
       right = left + 1;
@@ -152,43 +138,19 @@ class MinHeap {
 const minHeap = new MinHeap();
 const maxHeap = new MaxHeap();
 
-// arr.map((v, i) => {
-//   if (i % 2) minHeap.insert(v);
-//   else maxHeap.insert(v);
-//   if (minHeap.size && maxHeap.size && maxHeap.max() > minHeap.min()) {
-//     const left = maxHeap.max();
-//     const right = minHeap.min();
-//     maxHeap.delete();
-//     minHeap.delete();
-//     maxHeap.insert(right);
-//     minHeap.insert(left);
-//   }
-//   maxHeap.print();
-//   minHeap.print();
-//   answer.push(maxHeap.max());
-// });
-
-[3, 1, 5, 2, 4, -1, -9, 0, 7, -11, 5, 2, 4].map((v) => {
-  maxHeap.insert(v);
-  maxHeap.print();
+arr.map((v, i) => {
+  if (i % 2) minHeap.insert(v);
+  else maxHeap.insert(v);
+  if (minHeap.size && maxHeap.size && maxHeap.max() > minHeap.min()) {
+    const left = maxHeap.max();
+    const right = minHeap.min();
+    maxHeap.delete();
+    minHeap.delete();
+    maxHeap.insert(right);
+    minHeap.insert(left);
+  }
+  answer.push(maxHeap.max());
 });
-
-console.log("---------------------------------");
-for (let i = 0; i < 13; i++) {
-  maxHeap.delete();
-  maxHeap.print();
-}
-
-[3, 1, 5, 2, 4, -1, -9, 0, 7, -11, 5, 2, 4].map((v) => {
-  minHeap.insert(v);
-  minHeap.print();
-});
-
-console.log("---------------------------------");
-for (let i = 0; i < 13; i++) {
-  minHeap.delete();
-  minHeap.print();
-}
 
 console.log(answer.join("\n"));
